requirements.txt
pyTelegramBotAPI
numpy
scipy
librosa
unidecode
inflect
pydub
telebot
python-dotenv
matplotlib